[adjustment][pat] add modal styles
