+++ src/components/AddButton/styles.scss
@@ -20,7 +20,4 @@
       color: $secondary-background;
     }
   }
-  .modal-color {
-    color: cadetblue;
-  }
 }
