+++ src/utils/variables.scss
@@ -1,6 +1,7 @@
 $primary-background: #15322f;
 $secondary-background: #62a09d;
 $secondary-color: #8dc7c3;
+$modal-color: #ceebe9;
 $white-secondary: #ffff;
 $white-primary: #e0f0f5;
 $snippet-card: #e5e5e5;
